requirements.txt
-i https://pypi.org/simple
pyaes==1.6.1
pyasn1==0.4.8
python-decouple==3.3
rsa==4.6; python_version >= '3.5' and python_version < '4'
telethon==1.16.4
